[Fixes] Remainder is now an unsigned long int

+++ 0x08-palindrome_integer/0-is_palindrome.c
@@ -9,7 +9,7 @@
 int reversed(unsigned long n)
 {
 	unsigned long r = 0;
-	int remainder;
+	unsigned long remainder;
 
 	while (n)   /*Calculates the reversed number.*/
 	{
